refactor: :recycle: remove duplicated Django views and consolidated into `column_review.py`
